fix(server): graph route respects the route->service->db boundary
check-architecture-boundaries greps routes/ for direct-IO markers
(including the literal word 'qdrant', case-insensitive, even in doc
comments) and failed CI's changes-and-static gate. Two things fixed:
the DB orchestration (org transaction, edge queries, prune/degree/
community labelling) moves out of the route into
services::graph::build_org_graph so the handler is permission check +
service call + DTO mapping, matching every other route's layering; and
the route's comments now say 'vector index' instead of naming the
vendor, which is what the checker actually pattern-matches. The
community-label unit tests move with the function.
